backend/requirements.txt
fastapi
uvicorn
qdrant-client
requests
pydantic
python-multipart
pypdf
pymongo
sqlalchemy
pymysql
email-validator
cryptography
bcrypt
pyjwt
python-multipart
fastapi-mail
scikit-learn
numpy